+++ hc/front/forms.py
@@ -27,6 +27,10 @@ def _is_latin1(s: str) -> bool:
         return False
 
 
+def _choices(csv: str) -> list[tuple[str, str]]:
+    return [(v, v) for v in csv.split(",")]
+
+
 class LaxURLField(forms.URLField):
     default_validators = [WebhookValidator()]
 
@@ -85,9 +89,7 @@ class NameTagsForm(forms.Form):
 
 
 class AddCheckForm(NameTagsForm):
-    kind = forms.ChoiceField(
-        choices=(("simple", "simple"), ("cron", "cron"), ("oncalendar", "oncalendar"))
-    )
+    kind = forms.ChoiceField(choices=_choices("simple,cron,oncalendar"))
     timeout = forms.IntegerField(min_value=60, max_value=31536000)
     schedule = forms.CharField(required=False, max_length=100)
     tz = forms.CharField(max_length=36, validators=[TimezoneValidator()])
@@ -198,19 +200,16 @@ class AddUrlForm(forms.Form):
     value = LaxURLField(max_length=1000)
 
 
-METHODS = ("GET", "POST", "PUT")
-
-
 class WebhookForm(forms.Form):
     error_css_class = "has-error"
     name = forms.CharField(max_length=100, required=False)
 
-    method_down = forms.ChoiceField(initial="GET", choices=zip(METHODS, METHODS))
+    method_down = forms.ChoiceField(initial="GET", choices=_choices("GET,POST,PUT"))
     body_down = forms.CharField(max_length=1000, required=False)
     headers_down = HeadersField(required=False)
     url_down = LaxURLField(max_length=1000, required=False)
 
-    method_up = forms.ChoiceField(initial="GET", choices=zip(METHODS, METHODS))
+    method_up = forms.ChoiceField(initial="GET", choices=_choices("GET,POST,PUT"))
     body_up = forms.CharField(max_length=1000, required=False)
     headers_up = HeadersField(required=False)
     url_up = LaxURLField(max_length=1000, required=False)
@@ -404,3 +403,10 @@ class TransferForm(forms.Form):
 
 class AddTelegramForm(forms.Form):
     project = forms.UUIDField()
+
+
+class BadgeSettingsForm(forms.Form):
+    target = forms.ChoiceField(choices=_choices("all,tag"))
+    tag = forms.CharField(max_length=100, required=False)
+    fmt = forms.ChoiceField(choices=_choices("svg,json,shields"))
+    states = forms.ChoiceField(choices=_choices("2,3"))